+++ src/ft_print_decimal.c
@@ -6,7 +6,7 @@
 /*   By: hokutosuzuki <hosuzuki@student.42toky      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/30 17:23:15 by hokutosuz         #+#    #+#             */
-/*   Updated: 2022/02/10 13:05:10 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2022/02/10 15:33:16 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 void	ft_print_decimal(t_stock *lst, int decimal)
 {
-	size_t	len;
-	char		*res;
+	int		len;
+	char	*res;
 
 	res = ft_itoa(decimal);
 	len = ft_strlen(res);
